refactor: playerRegistryを分割
